Cash Register Function
This is a Cash Register function that accepts purchase price as the first argument, payment as the second argument, and cash-in-drawer (cid) as the third argument.
This script, when ran, will calculate the change that it should dispense to complete a transaction. It will then check to see if it can provide exact change from the Cash In Drawer (cid). If it can provide the change, it sorts the denominations to be returned in the highest-to-lowest order. If the array passed for "cid" does not have enough money to return the change, it will return "Insufficient Funds".